+++ Kernel/Output/HTML/LinkObject/Ticket.pm
@@ -307,7 +307,7 @@ sub TableCreateComplex {
     # Sort
     COLUMN:
     for my $Column ( sort { $SortOrder{$a} <=> $SortOrder{$b} } keys %UserColumns ) {
-        next COLUMN if $Column eq 'TicketNumber';    # Always present, already added.
+        next COLUMN if $Column eq 'TicketNumber';      # Always present, already added.
 
         # if enabled by default
         if ( $UserColumns{$Column} == 2 ) {
@@ -757,6 +757,15 @@ sub SearchOptionList {
             };
     }
 
+    if ( $Kernel::OM->Get('Kernel::Config')->Get('Ticket::ArchiveSystem') ) {
+        push @SearchOptionList,
+            {
+            Key  => 'ArchiveID',
+            Name => 'Archive search',
+            Type => 'List',
+            };
+    }
+
     # add formkey
     for my $Row (@SearchOptionList) {
         $Row->{FormKey} = 'SEARCH::' . $Row->{Key};
@@ -796,6 +805,8 @@ sub SearchOptionList {
             my @FormData = $Kernel::OM->Get('Kernel::System::Web::Request')->GetArray( Param => $Row->{FormKey} );
             $Row->{FormData} = \@FormData;
 
+            my $Multiple = 1;
+
             my %ListData;
             if ( $Row->{Key} eq 'StateIDs' ) {
                 %ListData = $Kernel::OM->Get('Kernel::System::State')->StateList(
@@ -812,6 +823,17 @@ sub SearchOptionList {
                     UserID => $Self->{UserID},
                 );
             }
+            elsif ( $Row->{Key} eq 'ArchiveID' ) {
+                %ListData = (
+                    ArchivedTickets    => Translatable('Archived tickets'),
+                    NotArchivedTickets => Translatable('Unarchived tickets'),
+                    AllTickets         => Translatable('All tickets'),
+                );
+                if ( !scalar @{ $Row->{FormData} } ) {
+                    $Row->{FormData} = ['NotArchivedTickets'];
+                }
+                $Multiple = 0;
+            }
 
             # add the input string
             $Row->{InputStrg} = $Self->{LayoutObject}->BuildSelection(
@@ -819,8 +841,32 @@ sub SearchOptionList {
                 Name       => $Row->{FormKey},
                 SelectedID => $Row->{FormData},
                 Size       => 3,
-                Multiple   => 1,
+                Multiple   => $Multiple,
                 Class      => 'Modernize',
+            );
+
+            next ROW;
+        }
+
+        if ( $Row->{Type} eq 'Checkbox' ) {
+
+            # get form data
+            $Row->{FormData} = $Kernel::OM->Get('Kernel::System::Web::Request')->GetParam( Param => $Row->{FormKey} );
+
+            # parse the input text block
+            $Self->{LayoutObject}->Block(
+                Name => 'Checkbox',
+                Data => {
+                    Name    => $Row->{FormKey},
+                    Title   => $Row->{FormKey},
+                    Content => $Row->{FormKey},
+                    Checked => $Row->{FormData} || '',
+                },
+            );
+
+            # add the input string
+            $Row->{InputStrg} = $Self->{LayoutObject}->Output(
+                TemplateFile => 'LinkObject',
             );
 
             next ROW;